% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{\%nin\%}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched}

\item{y}{vector or \code{NULL}: the values to be matched against}
}
\value{
A logical vector
}
\description{
Opposite of the \%in\% operator. Tests if X is NOT in Y.
}
\section{Date submitted}{
 2019-03-19
}

\section{Last Modified}{
 2019-03-19
}

\examples{
20 \%nin\% 1:10

}
\author{
David Wilkinson \email{davidpw@student.unimelb.edu.au}
}
