% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_fs.R
\name{print_fs}
\alias{print_fs}
\title{print_fs}
\usage{
print_fs(path = ".", depth = 2L)
}
\arguments{
\item{path}{file path to location where you want to visualise the file structure}

\item{depth}{an integer to set how deep you want to visualise the file structure}
}
\value{
Prints file structure to console
}
\description{
Print a file structure to the console. Akin to the Unix 'tree' command
}
\section{Date submitted}{
 2019-03-26
}

\section{Last Modified}{
 2019-03-26
}

\examples{
print_fs(depth = 2)

}
\author{
David Wilkinson \email{davidpw@student.unimelb.edu.au}
}
