% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spoof_legend.R
\name{spoof_legend}
\alias{spoof_legend}
\title{spoof_legend}
\usage{
spoof_legend(cs)
}
\arguments{
\item{colours}{a colour palette character string}

\item{labels}{a logical indicating if you want labels on the legend}

\item{label_text}{a character string of labels for each element in the colour palette. If NULL, defaults to colour names}

\item{borders}{a logical indicating if you want borders around each colour element}

\item{ncol}{a numeric indicating the number of columns for the legend}

\item{nrow}{a numeric indicating the number of rows for the legend}

\item{cex_label}{a numeric indicating the relative text size of the labels}

\item{label_padding_x}{a numeric indicating spacing of the labels on the x-axis relative to the starting point}

\item{label_padding_y}{a numeric indicating spacing of the labels on the y-axis relative to the starting point}

\item{...}{additional arguments can be passed to the text() function controlling the labels}
}
\description{
Make a fake legend using a colour palette
}
\section{Date submitted}{
 2019-10-15
}

\section{Last Modified}{
 2019-10-15
}

\examples{
spoof_legend(rainbow(10), ncol = 5, nrow = 2)

}
\author{
David Wilkinson \email{davidpw@student.unimelb.edu.au}
}
